Minor cleanup.  Move generator code to a separate file
